# docs/openapi.yaml
openapi: 3.0.3
info:
  title: API Asistencia Club MAEX
  description: |
    API REST del sistema de asistencia, evaluaciones e informes del Club Deportivo San José de Las Vegas (MAEX).

    ## Antes de probar un endpoint

    1. Obtenga un JWT con `POST /api/auth/login` o `POST /api/auth/microsoft/token`.
    2. En Swagger pulse **Authorize** y escriba: `Bearer <su_token>` (incluya la palabra Bearer).
    3. Consulte la **Guía de IDs** en el portal `/documentacion` o en `GUIA_IDENTIFICADORES.md`.

    ## ¿De dónde saco el ID del curso?

    Llame primero a **`GET /api/cursos`** con su correo (`?correo=entrenador@club.com`).
    En la respuesta, cada curso trae **`ID_Curso`** (ej. `"1847"`). Ese valor se usa como:
    - `idCurso` en `GET /api/inscritos`
    - `idcurso` en `POST /api/asistencia`

    ## Formato de respuestas (schemas)

    - **Éxito:** `{ "success": true, "message": "...", "data": { ... } }` → el payload útil está en **`data`**.
    - **Error:** `{ "success": false, "message": "..." }`

    Ver glosario completo en el portal de documentación (sección *Glosario de schemas*).

    ## Roles

    | Rol | Login |
    |-----|--------|
    | Entrenador / Administrador | Microsoft OAuth |
    | Proveedor / Desarrollador | Email + contraseña |

    Mantener este archivo sincronizado al modificar rutas en `src/routes/`.
  version: 1.0.0
  contact:
    name: SitiAutomate
servers:
  - url: http://localhost:3006
    description: Desarrollo local
  - url: '{appUrl}'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Token JWT de sesión (login o Microsoft)
    integracionBearer:
      type: http
      scheme: bearer
      description: Token fijo configurado en BEARERINS (integración externa)
  schemas:
...
    AsistenciaRegistro:
      type: object
      description: |
        Cuerpo de POST /api/asistencia. El responsable (email) se toma del JWT automáticamente.
        `idcurso` = mismo valor que `ID_Curso` de GET /api/cursos.
        `documento` = documento del participante de GET /api/inscritos.
      required: [documento, nombre, idcurso, curso, reporte]
      properties:
        documento:
          type: string
          example: "1012345678"
        nombre:
          type: string
          example: Juan Pérez
        idcurso:
          type: string
          example: "1847"
          description: Copiar de ID_Curso (GET /api/cursos).
        curso:
          type: string
          example: Fútbol categoría A
          description: Nombre legible del curso (texto libre, coherente con el id).
        reporte:
          type: string
          enum: [Asistió, Faltó, Excusa]
        comentarios:
          type: string
        ruta:
          type: string
        sede:
          type: string
    RubricaBody:
      type: object
      properties:
        nombre:
          type: string
        tipo:
          type: string
        descripcion:
          type: string
        categoria:
          type: string
        nombreCategoria:
          type: string
        alto:
          type: string
        medio:
          type: string
        bajo:
          type: string
        actividad:
          type: string
        comun:
          type: boolean
        estado:
          type: string
...
    CursoItem:
      type: object
      description: Elemento de `data.cursos[]` en GET /api/cursos. El identificador del curso es **ID_Curso**.
      properties:
        ID_Curso:
          type: string
          example: "1847"
          description: Usar como idCurso (inscritos) e idcurso (asistencia).
        Nombre_del_curso:
          type: string
          example: Fútbol categoría A
        Nombre_Corto_Curso:
          type: string
          example: FUT-A
        Linea:
          type: integer
          example: 3
        nombreLinea:
          type: string
          example: Deportes
        Actividad:
          type: string
    CursosListResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
        data:
          type: object
          properties:
            cursos:
              type: array
              items:
                $ref: '#/components/schemas/CursoItem'
            tieneApoyoGlobal:
              type: boolean
            esLider:
              type: boolean
            actividadesLider:
              type: array
              items:
                type: string
...
paths:
...
  /api/inscritos:
    get:
      tags: [Inscritos]
      summary: Listar inscritos
      security:
        - bearerAuth: []
      description: |
        Lista participantes inscritos en un curso.

        **Paso previo:** `GET /api/cursos?correo=...` y copie un valor de **`ID_Curso`**.
        Páselo aquí como **`idCurso`** (mismo string, ej. `1847`).

        El **documento del participante** en la respuesta se usa luego en asistencia (`documento`) y evaluaciones (`identificacion`).
      parameters:
        - name: idCurso
          in: query
          description: Valor `ID_Curso` devuelto por GET /api/cursos (no es el nombre del curso).
          schema:
            type: string
            example: "1847"
        - name: estado
          in: query
          description: Estados separados por coma (ej. CONFIRMADO,ACTIVO)
          schema:
            type: string
        - name: lite
          in: query
          schema:
            type: boolean
        - name: withRutaExtra
          in: query
          schema:
            type: boolean
      responses:
        '200':
          description: Lista de inscritos
  /api/cursos:
    get:
      tags: [Cursos]
      summary: Listar cursos activos
      security:
        - bearerAuth: []
      description: |
        **Endpoint principal para obtener el ID del curso.**

        - El correo del docente se toma del JWT salvo rol Administrador.
        - Para listar todos: `scope=all` (solo Administrador).
        - En `data.cursos[]` use el campo **`ID_Curso`** en llamadas posteriores.

        No invente IDs: deben existir en esta respuesta.
      parameters:
        - name: correo
          in: query
          required: false
          description: Email del docente (opcional; no-admin usa el correo del JWT).
          schema:
            type: string
            example: entrenador@club.com
        - name: soloMisCursos
          in: query
          schema:
            type: boolean
        - name: scope
          in: query
          description: Use `all` para ámbito global (administradores).
          schema:
            type: string
            enum: [all]
      responses:
        '200':
          description: Lista de cursos con ID_Curso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CursosListResponse'
              example:
                success: true
                message: OK
                data:
                  cursos:
                    - ID_Curso: "1847"
                      Nombre_del_curso: Fútbol categoría A
                      Nombre_Corto_Curso: FUT-A
                      Linea: 3
                      nombreLinea: Deportes
...
  /api/asistencia:
    get:
      tags: [Asistencia]
      summary: Consultar asistencia
      security:
        - bearerAuth: []
      description: Roles Administrador, Entrenador o Proveedor.
      parameters:
        - name: fechaInicio
          in: query
          schema:
            type: string
            format: date
        - name: fechaFin
          in: query
          schema:
            type: string
            format: date
        - name: q
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Registros de asistencia
        '401':
          description: No autenticado
        '403':
          description: Sin permisos
    post:
      tags: [Asistencia]
      summary: Registrar o actualizar asistencia del día
      description: |
        Registra o actualiza la asistencia de **hoy** para un participante en un curso.

        **Origen de datos:**
        1. `GET /api/cursos` → `idcurso` = `ID_Curso`
        2. `GET /api/inscritos?idCurso=...` → `documento` y `nombre`
        3. `reporte`: `Asistió` | `Faltó` | `Excusa`
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AsistenciaRegistro'
      responses:
        '200':
          description: Asistencia guardada
  /api/rubricas:
    get:
      tags: [Rúbricas]
      summary: Listar rúbricas
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Lista de rúbricas
    post:
      tags: [Rúbricas]
      summary: Crear rúbrica
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RubricaBody'
      responses:
        '201':
          description: Rúbrica creada
...
  /api/evaluaciones:
    post:
      tags: [Evaluaciones]
      summary: Crear o actualizar evaluación
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                foto:
                  type: string
                  format: binary
                participante:
                  type: string
                identificacion:
                  type: string
                categoria:
                  type: string
                comentario:
                  type: string
                rubricas:
                  type: string
                  description: JSON array de EvaluacionRubricaItem
                evaluacionId:
                  type: integer
                  description: Si se envía, actualiza evaluación existente
      responses:
        '200':
          description: Evaluación guardada
...